bench: refactor benchmarks to avoid macros and add extra comments
